Refactor auth hooks
