+++ frontend/src/components/PrivateRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '@/context/AuthContext';
+import { useAuth } from '@/context/useAuth';
 
 export default function PrivateRoute({ children }: { children: JSX.Element }) {
   const { isAuthenticated } = useAuth();
